Implemented most review comments
